Move all the toolbar creation function to a separate namespace+file.   Use the toolbar attribute of the panel in the index to setup the toolbar buttons and get rid of the toolbar logic in the panel.
