+++ src/tutorial.ts
@@ -1,35 +1,30 @@
-import {
-  Message
-} from '@phosphor/messaging';
+import { Message } from "@phosphor/messaging";
 
-import {
-  Widget, PanelLayout
-} from '@phosphor/widgets';
+import { Widget, PanelLayout } from "@phosphor/widgets";
 
-import '../style/index.css';
+import "../style/index.css";
 
 /**
  * An VoyagerTutorial viewer.
  */
-export
-class VoyagerTutorialWidget extends Widget {
+export class VoyagerTutorialWidget extends Widget {
   /**
    * Construct a new xkcd widget.
    */
   constructor(content: Widget) {
     super();
 
-    this.id = 'VoyagerTutorial-jupyterlab';
-    this.title.label = 'Tutorial:JupyterLab_Voyager';
+    this.id = "VoyagerTutorial-jupyterlab";
+    this.title.label = "Tutorial:JupyterLab_Voyager";
     this.title.closable = true;
-    this.addClass('jp-VoyagerTutorialWidget');
+    this.addClass("jp-VoyagerTutorialWidget");
     let toolbar = new Widget();
-    toolbar.addClass('jp-VoyagerTutorial-toolbar');  
-    let layout = this.layout = new PanelLayout();
+    toolbar.addClass("jp-VoyagerTutorial-toolbar");
+    let layout = (this.layout = new PanelLayout());
     layout.addWidget(toolbar);
     layout.addWidget(content);
-    
-   /*
+
+    /*
     this.img = document.createElement('img');
 
     let img0 = document.createElement('img');
@@ -192,17 +187,10 @@ class VoyagerTutorialWidget extends Widget {
                 </div>
              </body> 
                   `);*/
-    
   }
-
-  /**
-   * The image element associated with the widget.
-   */
-    readonly img: HTMLImageElement;
 
   /**
    * Handle update requests for the widget.
    */
-    onUpdateRequest(msg: Message): void {
-  }
-};
+  onUpdateRequest(msg: Message): void {}
+}
