+++ src/voyagerpanel.ts
@@ -39,44 +39,18 @@ import { read } from "vega-loader";
 import { PromiseDelegate } from "@phosphor/coreutils";
 
 import "../style/index.css";
+import { VoyagerToolbar } from "./voyagertoolbar";
 import { JupyterLab } from "@jupyterlab/application";
 
 /**
  * The mimetype used for Jupyter cell data.
  */
-const JUPYTER_CELL_MIME = "application/vnd.jupyter.cells";
-
-const VOYAGER_PANEL_TOOLBAR_CLASS = "jp-VoyagerPanel-toolbar";
-
-/**
- * The class name added to toolbar save button.
- */
-const TOOLBAR_SAVE_CLASS = "jp-SaveIcon";
-
-/**
- * The class name added to toolbar insert button.
- */
-const TOOLBAR_EXPORT_CLASS = "jp-ExportIcon";
-
-/**
- * The class name added to toolbar insert button.
- */
-const TOOLBAR_COPY_CLASS = "jp-CopyIcon";
-
-/**
- * The class name added to toolbar cut button.
- */
-const TOOLBAR_UNDO_CLASS = "jp-UndoIcon";
-
-/**
- * The class name added to toolbar copy button.
- */
-const TOOLBAR_REDO_CLASS = "jp-RedoIcon";
+export const JUPYTER_CELL_MIME = "application/vnd.jupyter.cells";
 
 /**
  * The class name added to a datavoyager widget.
  */
-const Voyager_CLASS = "jp-Voyager";
+export const Voyager_CLASS = "jp-Voyager";
 
 export class VoyagerPanel extends DocumentWidget<Widget> {
   public voyager_cur!: Voyager;
@@ -94,8 +68,14 @@ export class VoyagerPanel extends DocumentWidget<Widget> {
 
     this.title.label = PathExt.basename(this.context.path);
 
-    this.context.model.contentChanged.connect(this.update, this);
-    this.context.fileChanged.connect(this.update, this);
+    this.context.model.contentChanged.connect(
+      this.update,
+      this
+    );
+    this.context.fileChanged.connect(
+      this.update,
+      this
+    );
 
     this.context.ready.then(() => {
       const data = this.context.model.toString();
@@ -190,18 +170,7 @@ export class VoyagerPanel extends DocumentWidget<Widget> {
     });
 
     // Toolbar
-    this.toolbar.addClass(VOYAGER_PANEL_TOOLBAR_CLASS);
-    this.toolbar.addItem("save", Private.createSaveButton(this));
-    this.toolbar.addItem(
-      "saveAs",
-      Private.createExportButton(this, app, docManager)
-    );
-    this.toolbar.addItem(
-      "ExportToNotebook",
-      Private.createCopyButton(this, app, docManager)
-    );
-    this.toolbar.addItem("undo", Private.createUndoButton(this));
-    this.toolbar.addItem("redo", Private.createRedoButton(this));
+    VoyagerToolbar.setupToolbar(this.toolbar, this, app, docManager);
   }
 
   /**
@@ -216,7 +185,10 @@ export class VoyagerPanel extends DocumentWidget<Widget> {
     }
     this._settings = settings;
     if (this._settings) {
-      this._settings.changed.connect(this._onSettingsChanged, this);
+      this._settings.changed.connect(
+        this._onSettingsChanged,
+        this
+      );
     }
     this.update();
   }
@@ -327,18 +299,7 @@ export class VoyagerPanel_DF extends DocumentWidget<Widget> {
     });
 
     // Toolbar
-    this.toolbar.addClass(VOYAGER_PANEL_TOOLBAR_CLASS);
-    this.toolbar.addItem("save", Private.createSaveButton(this));
-    this.toolbar.addItem(
-      "saveAs",
-      Private.createExportButton(this, app, docManager)
-    );
-    this.toolbar.addItem(
-      "ExportToNotebook",
-      Private.createCopyButton(this, app, docManager)
-    );
-    this.toolbar.addItem("undo", Private.createUndoButton(this));
-    this.toolbar.addItem("redo", Private.createRedoButton(this));
+    VoyagerToolbar.setupToolbar(this.toolbar, this, app, docManager);
   }
 
   /**
@@ -365,213 +326,6 @@ namespace Private {
     relatedViews: "initiallyCollapsed",
     wildcards: "enabled"
   };
-
-  export function createSaveButton(
-    widget: VoyagerPanel | VoyagerPanel_DF
-  ): ToolbarButton {
-    return new ToolbarButton({
-      className: TOOLBAR_SAVE_CLASS,
-      onClick: () => {
-        if (
-          widget &&
-          widget.hasClass(Voyager_CLASS) &&
-          (widget as VoyagerPanel).context.path.indexOf("vl.json") !== -1
-        ) {
-          var datavoyager = (widget as VoyagerPanel).voyager_cur;
-          var dataSrc = (widget as VoyagerPanel).data_src;
-          let spec = datavoyager.getSpec(false);
-          let context = widget.context as Context<DocumentRegistry.IModel>;
-          context.model.fromJSON({
-            data: dataSrc,
-            mark: spec.mark,
-            encoding: spec.encoding,
-            height: spec.height,
-            width: spec.width,
-            description: spec.description,
-            name: spec.name,
-            selection: spec.selection,
-            title: spec.title,
-            transform: spec.transform
-          });
-          context.save();
-        } else {
-          showDialog({
-            title: "Source File Type is NOT Vega-Lite (.vl.json)",
-            body: "To save this chart, use 'Export Voyager as Vega-Lite file' ",
-            buttons: [Dialog.warnButton({ label: "OK" })]
-          });
-        }
-      },
-      tooltip: "Save Voyager"
-    });
-  }
-
-  export function createExportButton(
-    widget: VoyagerPanel | VoyagerPanel_DF,
-    app: JupyterLab,
-    docManager: DocumentManager
-  ): ToolbarButton {
-    return new ToolbarButton({
-      className: TOOLBAR_EXPORT_CLASS,
-      onClick: () => {
-        var datavoyager = (widget as VoyagerPanel).voyager_cur;
-        var dataSrc = (widget as VoyagerPanel).data_src;
-        //let aps = datavoyager.getApplicationState();
-        let spec = datavoyager.getSpec(false);
-        //let context = docManager.contextForWidget(widget) as Context<DocumentRegistry.IModel>;
-        let context = widget.context as Context<DocumentRegistry.IModel>;
-        let path = PathExt.dirname(context.path);
-        var content: any;
-        if (spec !== undefined) {
-          content = {
-            data: dataSrc,
-            mark: spec.mark,
-            encoding: spec.encoding,
-            height: spec.height,
-            width: spec.width,
-            description: spec.description,
-            name: spec.name,
-            selection: spec.selection,
-            title: spec.title,
-            transform: spec.transform
-          };
-        } else {
-          content = {
-            data: dataSrc
-          };
-        }
-        let input_block = document.createElement("div");
-        let input_prompt = document.createElement("div");
-        input_prompt.textContent = "";
-        let input = document.createElement("input");
-        input_block.appendChild(input_prompt);
-        input_block.appendChild(input);
-        let bd = new Widget({ node: input_block });
-        showDialog({
-          title: "Export as Vega-Lite File (.vl.json)",
-          body: bd,
-          buttons: [Dialog.cancelButton(), Dialog.okButton({ label: "OK" })]
-        }).then(result => {
-          let msg = input.value;
-          if (result.button.accept) {
-            if (!isValidFileName(msg)) {
-              showErrorMessage(
-                "Name Error",
-                Error(
-                  `"${result.value}" is not a valid name for a file. ` +
-                    `Names must have nonzero length, ` +
-                    `and cannot include "/", "\\", or ":"`
-                )
-              );
-            } else {
-              let basePath = path;
-              let newPath = PathExt.join(
-                basePath,
-                msg.indexOf(".vl.json") !== -1 ? msg : msg + ".vl.json"
-              );
-              app.commands
-                .execute("docmanager:new-untitled", {
-                  path: basePath,
-                  ext: ".vl.json",
-                  type: "file"
-                })
-                .then(model => {
-                  docManager.rename(model.path, newPath).then(model => {
-                    app.commands
-                      .execute("docmanager:open", {
-                        path: model.path,
-                        factory: "Editor"
-                      })
-                      .then(widget => {
-                        let context = docManager.contextForWidget(widget);
-                        if (context != undefined) {
-                          context.save().then(() => {
-                            if (context != undefined) {
-                              context.model.fromJSON(content);
-                              context.save();
-                            }
-                          });
-                        }
-                      });
-                  });
-                });
-            }
-          }
-        });
-      },
-      tooltip: "Export Voyager as Vega-Lite File"
-    });
-  }
-
-  export function createCopyButton(
-    widget: VoyagerPanel | VoyagerPanel_DF,
-    app: JupyterLab,
-    docManager: DocumentManager
-  ): ToolbarButton {
-    return new ToolbarButton({
-      className: TOOLBAR_COPY_CLASS,
-      onClick: () => {
-        var datavoyager = widget.voyager_cur;
-        var dataSrc = widget.data_src;
-        let spec = datavoyager.getSpec(false);
-        let src = JSON.stringify({
-          data: dataSrc,
-          mark: spec.mark,
-          encoding: spec.encoding,
-          height: spec.height,
-          width: spec.width,
-          description: spec.description,
-          name: spec.name,
-          selection: spec.selection,
-          title: spec.title,
-          transform: spec.transform
-        });
-        let clipboard = Clipboard.getInstance();
-        clipboard.clear();
-        let data = [
-          {
-            cell_type: "code",
-            execution_count: null,
-            metadata: {},
-            outputs: [],
-            source: [
-              "import altair as alt\n",
-              "import pandas as pd\n",
-              "import json\n",
-              `data_src = json.loads('''${src}''')\n`,
-              "alt.Chart.from_dict(data_src)\n"
-            ]
-          }
-        ];
-        clipboard.setData(JUPYTER_CELL_MIME, data);
-      },
-      tooltip: "Copy Altair Graph to clipboard"
-    });
-  }
-
-  export function createUndoButton(
-    widget: VoyagerPanel | VoyagerPanel_DF
-  ): ToolbarButton {
-    return new ToolbarButton({
-      className: TOOLBAR_UNDO_CLASS,
-      onClick: () => {
-        (widget as VoyagerPanel).voyager_cur.undo();
-      },
-      tooltip: "Undo"
-    });
-  }
-
-  export function createRedoButton(
-    widget: VoyagerPanel | VoyagerPanel_DF
-  ): ToolbarButton {
-    return new ToolbarButton({
-      className: TOOLBAR_REDO_CLASS,
-      onClick: () => {
-        (widget as VoyagerPanel).voyager_cur.redo();
-      },
-      tooltip: "Redo"
-    });
-  }
 
   export function isValidURL(str: string) {
     var a = document.createElement("a");
